requirements.txt
python-telegram-bot==21.10
openai>=1.60.0
httpx>=0.27.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
duckduckgo-search>=7.0.0
beautifulsoup4>=4.12.0
